+++ src/components/progress-ring/progress-ring.e2e.ts
@@ -1,16 +1,16 @@
-import { newE2EPage } from '@stencil/core/testing';
+import { newE2EPage } from "@stencil/core/testing";
 
-describe('progress-ring', () => {
-  it('should render', async () => {
+describe("progress-ring", () => {
+  it("should render", async () => {
     const page = await newE2EPage();
-    await page.setContent('<progress-ring></progress-ring>');
-    const element = await page.find('progress-ring');
+    await page.setContent("<progress-ring></progress-ring>");
+    const element = await page.find("progress-ring");
     expect(element).not.toBeNull();
   });
-  it('should render textContent', async () => {
+  it("should render textContent", async () => {
     const page = await newE2EPage();
     await page.setContent('<progress-ring percentage="30"></progress-ring>');
-    const element = await page.find('progress-ring >>> .percentageText');
-    expect(element.textContent).toEqualText('%');
+    const element = await page.find("progress-ring >>> .percentageText");
+    expect(element.textContent).toEqualText("%");
   });
 });