+++ CHANGELOG.md
@@ -1,63 +1,87 @@
 # Changelog
 
 ## 0.0.7 (2019-02-28)
+
 - Fixed vertical alignment issue of the digits
 - Added font style to examples
 
 ## 0.0.8 (2019-02-28)
+
 - Updated inline demo style
 
 ## 0.0.9 (2019-03-01)
+
 - Fixed invertColors and disableDigits issues
 
 ## 0.0.10 (2019-03-02)
+
 - Updated integration guide on readme
 
 ## 0.0.11 (2019-03-02)
+
 - Fixed typo on readme
 
 ## 0.0.12 (2019-03-03)
+
 - Added CSS transition effects
 
 ## 0.0.13 (2019-09-05)
+
 - Updated dependencies
 - Made enhancement to the transitions
 
 ## 1.0.0 (2019-09-05)
+
 - Fixed a transition issue with restart before completing the initial animation loop
 
 ## 1.0.1 (2019-09-11)
+
 - Fixed decimals issue for zero percent
 
 ## 1.0.2 (2019-09-26)
+
 - Added missing implementation of restartProgress in several places
 
 ## 1.0.3 (2020-05-18)
+
 - Updated dependencies
 
 ## 1.0.4 (2021-06-10)
+
 - Upgraded Stencil version to 2.6
 
 ## 1.0.5 (2021-06-12)
+
 - Fixed animations not resuming after getting reattached to DOM
 
 ## 1.0.6 (2021-11-08)
+
 - Upgraded Stencil version to 2.10
 - Disabling e2e test until stencil supports jest v27
 - Updated README to include framework integration guide
 
 ## 1.0.7 (2021-11-12)
+
 - Changed percent prop to percentage prop
 - Added roundLinecap prop
 - Removed decimal digits for 0 percent
 - Integrated slotted elements
 
 ## 1.0.8 (2021-11-12)
+
 - Fixed zero percent issue (digits gets hidden)
 - Added disable-decimals
 
 ## 1.0.9 (2021-11-12)
+
 - Fixed broken demo
 
 ## 1.0.10 (2021-11-16)
+
 - Updated easing-animation-frames version to 1.0.2
+
+## 1.0.11 (2021-11-18)
+
+- Added callback functions that emit events
+- Updated eslint and prettier settings
+- Updated easing-animation-frames version to 1.0.3